+++ RSwitch/Swift/RStudioServerSessionWebViewController.swift
@@ -51,9 +51,6 @@ class RstudioServerSessionWebViewController: NSViewController, NSWindowDelegate 
 extension RstudioServerSessionWebViewController: WKUIDelegate {
 
   func webView(_ webView: WKWebView, runOpenPanelWith parameters: WKOpenPanelParameters, initiatedByFrame frame: WKFrameInfo, completionHandler: @escaping ([URL]?) -> Void) {
-    
-    NSLog("PANELING!")
-    print("PANELING!")
 
     let openPanel = NSOpenPanel()
     
@@ -79,9 +76,7 @@ extension RstudioServerSessionWebViewController: WKUIDelegate {
       let u = URL(string: navigationAction.request.url!.absoluteString)!
       
       if ((u.urlComponents?.path.starts(with: "/export/")) != nil) {
-        
-        NSLog("Navigation action thing: " + navigationAction.request.url!.absoluteString)
-        
+                
         let exportWindowController = NSStoryboard(name: "Main", bundle: nil).instantiateController(withIdentifier: "exportPopupPanel") as! ExportWebViewController
         
         let exportWV = (exportWindowController.contentViewController as! exportPopupViewController)
@@ -89,42 +84,25 @@ extension RstudioServerSessionWebViewController: WKUIDelegate {
         exportWV.view.window?.title = navigationAction.request.url!.absoluteString
         
         exportWV.setupWebView(configuration: configuration)
-        //exportWindowController.showWindow(self)
-        
-        NSLog("Before exportWV.loadWebView")
-        
-//        exportWV.loadWebView(urlIn: navigationAction.request.url!.absoluteString)
+                
         exportWV.loadWebView(urlIn: "")
                 
-        NSLog("After exportWV.loadWebView")
-
         let urlPath = navigationAction.request.url!.absoluteString
-        
-        NSLog("loadWebView: \(urlPath)")
-        
+                
         // Check for "/export/"
         // If export, then get bring up a Save Panel and then download the file to that location
 
         if let url = URL(string: urlPath) {
-          
-          NSLog("URL path: \(url.path)")
-          
+                    
           if (url.path.starts(with: "/export")) {
-            
-            NSLog("  Name: " + url.queryParameters["name"]!)
-            
+                        
             let savePanel = NSSavePanel()
             
             savePanel.canCreateDirectories = true
             savePanel.nameFieldStringValue = url.queryParameters["name"]!
             savePanel.beginSheetModal(for:self.view.window!) { (response) in
-              if (response == NSApplication.ModalResponse.OK) {
-                            
+              if (response == NSApplication.ModalResponse.OK) {                            
                 download_from_studio_server(fromRS: url.absoluteString, toFS: savePanel.url!.absoluteString)
-                
-              } else {
-                
-                NSLog("Don't do anything!")
               }
               savePanel.close()
             }